requirements.txt
rembg[cpu]
pillow
click
tqdm
